Graphics.Vty.Attributes

Description

Display attributes

Attributes have three components: a foreground color, a background color, and a style mask. The simplest attribute is the default attribute, or defAttr. Attributes can be modified with withForeColor, withBackColor, and withStyle, e.g.,

    defAttr `withForeColor` red

Image constructors often require an Attr to indicate the attributes used in the image, e.g.,

    string (defAttr `withForeColor` red) "this text will be red"

The appearance of Images using defAttr is determined by the The terminal, so this is not something VTY can control. The user is free to The define the color scheme of the terminal as they see fit.

The value currentAttr will keep the attributes of whatever was output previously.

Documentation

module Graphics.Vty.Attributes.Color

data Attr

A display attribute defines the Color and Style of all the characters rendered after the attribute is applied.

At most 256 colors, picked from a 240 and 16 color palette, are possible for the background and foreground. The 240 colors and 16 colors are points in different palettes. See Color for more information.

Constructors

Attr

Fields

attrStyle :: !(MaybeDefault Style)

attrForeColor :: !(MaybeDefault Color)

attrBackColor :: !(MaybeDefault Color)

attrURL :: !(MaybeDefault Text)

data FixedAttr

Specifies the display attributes such that the final style and color values do not depend on the previously applied display attribute. The display attributes can still depend on the terminal's default colors (unfortunately).

Constructors

FixedAttr

Fields

fixedStyle :: !Style

fixedForeColor :: !(Maybe Color)

fixedBackColor :: !(Maybe Color)

fixedURL :: !(Maybe Text)

data MaybeDefault v where

The style and color attributes can either be the terminal defaults. Or be equivalent to the previously applied style. Or be a specific value.

Constructors

Default :: MaybeDefault v

KeepCurrent :: MaybeDefault v

SetTo :: forall v. (Eq v, Show v, Read v) => !v -> MaybeDefault v

defAttr :: Attr

Sets the style, background color and foreground color to the default values for the terminal. There is no easy way to determine what the default background and foreground colors are.

currentAttr :: Attr

Keeps the style, background color and foreground color that was previously set. Used to override some part of the previous style.

EG: current_style withForeColor brightMagenta

Would be the currently applied style (be it underline, bold, etc) but with the foreground color set to brightMagenta.

Styles

type Style = Word8

Styles are represented as an 8 bit word. Each bit in the word is 1 if the style attribute assigned to that bit should be applied and 0 if the style attribute should not be applied.

withStyle :: Attr -> Style -> Attr

Add the given style attribute

standout :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

italic :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

strikethrough :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

underline :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

reverseVideo :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

blink :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

dim :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

bold :: Style

Valid style attributes include:

standout

underline

reverseVideo

blink

dim

bold/bright

italic

strikethrough (via the smxx/rmxx terminfo capabilities)

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

defaultStyleMask :: Style

styleMask :: Attr -> Word8

hasStyle :: Style -> Style -> Bool

true if the given Style value has the specified Style set.

Setting attribute colors

withForeColor :: Attr -> Color -> Attr

Set the foreground color of an Attr.

withBackColor :: Attr -> Color -> Attr

Set the background color of an Attr.

Setting hyperlinks

withURL :: Attr -> Text -> Attr

Add a hyperlinked URL using the proposed escape sequences for hyperlinked URLs. These escape sequences are comparatively new and aren't widely supported in terminal emulators yet, but most terminal emulators that don't know about these sequences will ignore these sequences, and therefore this should fall back sensibly. In some cases they won't and this will result in garbage, so this is why hyperlinking is disabled by default, in which case this combinator has no observable effect. To enable it, enable Hyperlink mode on your Vty output interface.