Graphics.Vty.DisplayAttributes

Documentation

fixDisplayAttr :: FixedAttr -> Attr -> FixedAttr

Given the previously applied display attributes as a FixedAttr and the current display attributes as an Attr produces a FixedAttr that represents the current display attributes. This is done by using the previously applied display attributes to remove the KeepCurrent abstraction.

data DisplayAttrDiff

difference between two display attributes. Used in the calculation of the operations required to go from one display attribute to the next.

Previously, vty would reset display attributes to default then apply the new display attributes. This turned out to be very expensive: A *lot* more data would be sent to the terminal than required.

Constructors

DisplayAttrDiff

Fields

styleDiffs :: [StyleStateChange]

foreColorDiff :: DisplayColorDiff

backColorDiff :: DisplayColorDiff

urlDiff :: URLDiff

simplifyStyleDiffs :: [StyleStateChange] -> [StyleStateChange] -> [StyleStateChange]

Used in the computation of a final style attribute change.

simplifyColorDiffs :: DisplayColorDiff -> DisplayColorDiff -> DisplayColorDiff

Consider two display color attributes diffs. What display color attribute diff are these equivalent to?

simplifyUrlDiffs :: URLDiff -> URLDiff -> URLDiff

Consider two URL changes, which are mostly going to be the latter unless the latter specifies no change.

data DisplayColorDiff

Difference between two display color attribute changes.

Constructors

ColorToDefault

NoColorChange

SetColor !Color

data StyleStateChange

Style attribute changes are transformed into a sequence of apply/removes of the individual attributes.

Constructors

ApplyStandout

RemoveStandout

ApplyItalic

RemoveItalic

ApplyStrikethrough

RemoveStrikethrough

ApplyUnderline

RemoveUnderline

ApplyReverseVideo

RemoveReverseVideo

ApplyBlink

RemoveBlink

ApplyDim

RemoveDim

ApplyBold

RemoveBold

data URLDiff

Constructors

LinkTo !ByteString

NoLinkChange

EndLink

displayAttrDiffs :: FixedAttr -> FixedAttr -> DisplayAttrDiff

Determines the diff between two display&color attributes. This diff determines the operations that actually get output to the terminal.

diffURL :: Maybe Text -> Maybe Text -> URLDiff

diffColor :: Maybe Color -> Maybe Color -> DisplayColorDiff

diffStyles :: Style -> Style -> [StyleStateChange]