Graphics.Vty.Input.Focus

Documentation

requestFocusEvents :: String

These sequences set xterm-based terminals to send focus event sequences.

disableFocusEvents :: String

These sequences disable focus events.

isFocusEvent :: String -> Bool

Does the specified string begin with a focus event?

classifyFocusEvent :: String -> KClass

Attempt to classify an input string as a focus event.