Graphics.Vty.Input.Mouse

Description

This module provides parsers for mouse events for both "normal" and "extended" modes. This implementation was informed by

http://invisible-island.net/xterm/ctlseqs/ctlseqs.html#h2-Mouse-Tracking

Documentation

requestMouseEvents :: String

These sequences set xterm-based terminals to send mouse event sequences.

disableMouseEvents :: String

These sequences disable mouse events.

isMouseEvent :: String -> Bool

Does the specified string begin with a mouse event?

classifyMouseEvent :: String -> KClass

Attempt to lassify an input string as a mouse event.