Graphics.Vty.Output.Interface

Description

This module provides an abstract interface for performing terminal output. The only user-facing part of this API is Output.

Documentation

data Output

The Vty terminal output interface.

Constructors

Output

Fields

terminalID :: String

Text identifier for the output device. Used for debugging.

releaseTerminal :: IO ()

Release the terminal just prior to application exit and reset it to its state prior to application startup.

reserveDisplay :: IO ()

Clear the display and initialize the terminal to some initial display state.

The expectation of a program is that the display starts in some The initial state. initial state would consist of fixed values:

cursor at top left

UTF-8 character encoding

drawing characteristics are the default

releaseDisplay :: IO ()

Return the display to the state before reserveDisplay If no previous state then set the display state to the initial state.

setDisplayBounds :: (Int, Int) -> IO ()

Sets the current display bounds (width, height).

displayBounds :: IO DisplayRegion

Returns the current display bounds.

outputByteBuffer :: ByteString -> IO ()

Output the bytestring to the terminal device.

contextColorCount :: Int

Specifies the maximum number of colors supported by the context.

supportsCursorVisibility :: Bool

Specifies whether the cursor can be shown / hidden.

supportsMode :: Mode -> Bool

Indicates support for terminal modes for this output device.

setMode :: Mode -> Bool -> IO ()

Enables or disables a mode (does nothing if the mode is unsupported).

getModeStatus :: Mode -> IO Bool

Returns whether a mode is enabled.

assumedStateRef :: IORef AssumedState

mkDisplayContext :: Output -> DisplayRegion -> IO DisplayContext

Acquire display access to the given region of the display. Currently all regions have the upper left corner of (0,0) and the lower right corner at (max displayWidth providedWidth, max displayHeight providedHeight)

ringTerminalBell :: IO ()

Ring the terminal bell if supported.

supportsBell :: IO Bool

Returns whether the terminal has an audio bell feature.

supportsItalics :: IO Bool

Returns whether the terminal supports italicized text.

This is terminal-dependent and should make a best effort to determine whether this feature is supported, but even if the terminal advertises support (e.g. via terminfo) that might not be a reliable indicator of whether the feature will work as desired.

supportsStrikethrough :: IO Bool

Returns whether the terminal supports strikethrough text.

This is terminal-dependent and should make a best effort to determine whether this feature is supported, but even if the terminal advertises support (e.g. via terminfo) that might not be a reliable indicator of whether the feature will work as desired.

data AssumedState

Constructors

AssumedState

Fields

prevFattr :: Maybe FixedAttr

prevOutputOps :: Maybe DisplayOps

data DisplayContext

Constructors

DisplayContext

Fields

contextDevice :: Output

contextRegion :: DisplayRegion

Provide the bounds of the display context.

writeMoveCursor :: Int -> Int -> Write

Sets the output position to the specified row and column where the number of bytes required for the control codes can be specified seperate from the actual byte sequence.

writeShowCursor :: Write

writeHideCursor :: Write

writeSetAttr :: Bool -> FixedAttr -> Attr -> DisplayAttrDiff -> Write

writeDefaultAttr :: Bool -> Write

Reset the display attributes to the default display attributes.

writeRowEnd :: Write

inlineHack :: IO ()

See inlineHack

data Mode

Modal terminal features that can be enabled and disabled.

Constructors

Mouse
Mouse mode (whether the terminal is configured to provide mouse input events)

BracketedPaste
Paste mode (whether the terminal is configured to provide events on OS pastes)

Focus
Focus-in/focus-out events (whether the terminal is configured to provide events on focus change)

Hyperlink
Hyperlink mode via the withURL attribute modifier (see https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda). Note that this may not work gracefully in all terminal emulators so be sure to test this mode with the terminals you intend to support. It is off by default.

displayContext :: Output -> DisplayRegion -> IO DisplayContext

outputPicture :: DisplayContext -> Picture -> IO ()

Displays the given Picture.

The image is cropped to the display size.

Converted into a sequence of attribute changes and text spans.

The cursor is hidden.

Serialized to the display.

The cursor is then shown and positioned or kept hidden.

initialAssumedState :: AssumedState

limitAttrForDisplay :: Output -> Attr -> Attr

Not all terminals support all display attributes. This filters a display attribute to what the given terminal can display.