Graphics.Vty.UnicodeWidthTable.IO

Documentation

readUnicodeWidthTable :: FilePath -> IO (Either String UnicodeWidthTable)

Load a binary unicode width table from the specified file.

This either returns a successfully parsed table or a table parsing error message. This does not handle I/O exceptions.

parseUnicodeWidthTable :: ByteString -> Either String UnicodeWidthTable

Parse a binary unicode width table.

writeUnicodeWidthTable :: FilePath -> UnicodeWidthTable -> IO ()

Write the unicode width table to the specified path.

This does not handle I/O exceptions.