Graphics.Vty.UnicodeWidthTable.Types

Documentation

data UnicodeWidthTable

A run-length-encoded table of Unicode character widths.

Constructors

UnicodeWidthTable

Fields

unicodeWidthTableRanges :: [WidthTableRange]

The ranges in the table.

data WidthTableRange

A range of code points in a width table.

Constructors

WidthTableRange

Fields

rangeStart :: Word32

The range's starting code point.

rangeSize :: Word32

The number of code points in the contiguous range, beginning with the starting code point (rangeStart).

rangeColumns :: Word8

The terminal width, in columns, of all of the characters corresponding to the code points in this range.