Graphics.Vty.Attributes

Description

Display attributes

Attributes have three components: a foreground color, a background color, and a style mask. The simplest attribute is the default attribute, or defAttr. Attributes can be modified with withForeColor, withBackColor, and withStyle, e.g.,

    defAttr withForeColor red

Image constructors often require an Attr to indicate the attributes used in the image, e.g.,

    string (defAttr withForeColor red) "this text will be red"

The appearance of Images using defAttr is determined by the The terminal, so this is not something VTY can control. The user is free to The define the color scheme of the terminal as they see fit.

The value currentAttr will keep the attributes of whatever was output previously.

Documentation

data Attr

A display attribute defines the Color and Style of all the characters rendered after the attribute is applied.

At most 256 colors, picked from a 240 and 16 color palette, are possible for the background and foreground. The 240 colors and 16 colors are points in different palettes. See Color for more information.

Constructors

Attr

Fields

attrStyle :: !(MaybeDefault Style)

attrForeColor :: !(MaybeDefault Color)

attrBackColor :: !(MaybeDefault Color)

data FixedAttr

Specifies the display attributes such that the final style and color values do not depend on the previously applied display attribute. The display attributes can still depend on the terminal's default colors (unfortunately).

Constructors

FixedAttr

Fields

fixedStyle :: !Style

fixedForeColor :: !(Maybe Color)

fixedBackColor :: !(Maybe Color)

data MaybeDefault v where

The style and color attributes can either be the terminal defaults. Or be equivalent to the previously applied style. Or be a specific value.

Constructors

Default :: MaybeDefault v

KeepCurrent :: MaybeDefault v

SetTo :: forall v. (Eq v, Show v, Read v) => !v -> MaybeDefault v

defAttr :: Attr

Sets the style, background color and foreground color to the default values for the terminal. There is no easy way to determine what the default background and foreground colors are.

currentAttr :: Attr

Keeps the style, background color and foreground color that was previously set. Used to override some part of the previous style.

EG: current_style withForeColor brightMagenta

Would be the currently applied style (be it underline, bold, etc) but with the foreground color set to brightMagenta.

Styles

type Style = Word8

Styles are represented as an 8 bit word. Each bit in the word is 1 if the style attribute assigned to that bit should be applied and 0 if the style attribute should not be applied.

withStyle :: Attr -> Style -> Attr

Add the given style attribute

standout :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

underline :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

reverseVideo :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

blink :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

dim :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

bold :: Style

The 6 possible style attributes:

standout

underline

reverseVideo

blink

dim

bold/bright

(The invisible, protect, and altcharset display attributes some terminals support are not supported via VTY.)

defaultStyleMask :: Style

styleMask :: Attr -> Word8

hasStyle :: Style -> Style -> Bool

true if the given Style value has the specified Style set.

Setting attribute colors

withForeColor :: Attr -> Color -> Attr

Set the foreground color of an Attr.

withBackColor :: Attr -> Color -> Attr

Set the background color of an Attr.

Colors

data Color

Abstract data type representing a color.

Currently the foreground and background color are specified as points in either a:

16 color palette. Where the first 8 colors are equal to the 8 colors of the ISO 6429 (ANSI) 8 color palette and the second 8 colors are bright/vivid versions of the first 8 colors.

240 color palette. This palette is a regular sampling of the full RGB colorspace for the first 224 colors. The remaining 16 colors is a greyscale palette.

The 8 ISO 6429 (ANSI) colors are as follows:

black

red

green

yellow

blue

magenta

cyan

white

The mapping from points in the 240 color palette to colors actually displayable by the terminal depends on the number of colors the terminal claims to support. Which is usually determined by the terminfo "colors" property. If this property is not being accurately reported then the color reproduction will be incorrect.

If the terminal reports <= 16 colors then the 240 color palette points are only mapped to the 8 color pallete. I'm not sure of the RGB points for the "bright" colors which is why they are not addressable via the 240 color palette.

If the terminal reports > 16 colors then the 240 color palette points are mapped to the nearest points in a ("color count" - 16) subsampling of the 240 color palette.

All of this assumes the terminals are behaving similarly to xterm and rxvt when handling colors. And that the individual colors have not been remapped by the user. There may be a way to verify this through terminfo but I don't know it.

Seriously, terminal color support is INSANE.

Constructors

ISOColor !Word8

Color240 !Word8

black :: Color

Standard 8-color ANSI terminal color codes.

red :: Color

Standard 8-color ANSI terminal color codes.

green :: Color

Standard 8-color ANSI terminal color codes.

yellow :: Color

Standard 8-color ANSI terminal color codes.

blue :: Color

Standard 8-color ANSI terminal color codes.

magenta :: Color

Standard 8-color ANSI terminal color codes.

cyan :: Color

Standard 8-color ANSI terminal color codes.

white :: Color

Standard 8-color ANSI terminal color codes.

brightBlack :: Color

Bright/Vivid variants of the standard 8-color ANSI

brightRed :: Color

Bright/Vivid variants of the standard 8-color ANSI

brightGreen :: Color

Bright/Vivid variants of the standard 8-color ANSI

brightYellow :: Color

Bright/Vivid variants of the standard 8-color ANSI

brightBlue :: Color

brightMagenta :: Color

brightCyan :: Color

brightWhite :: Color

rgbColor :: Integral i => i -> i -> i -> Color

Create a Vty Color (in the 240 color set) from an RGB triple.