Graphics.Vty.Input.Paste

Description

This module provides bracketed paste support as described at

http://cirw.in/blog/bracketed-paste

Documentation

parseBracketedPaste :: String -> KClass

Parse a bracketed paste. This should only be called on a string if both bracketedPasteStarted and bracketedPasteFinished return True.

bracketedPasteStarted :: String -> Bool

Does the input start a bracketed paste?

bracketedPasteFinished :: String -> Bool

Does the input contain a complete bracketed paste?